+++ language-server/src/tests/hover.rs
@@ -1949,3 +1949,52 @@ type Wibble {
         find_position_of("Wibble")
     );
 }
+
+#[test]
+fn hover_type_constructor_with_no_fields() {
+    assert_hover!(
+        "
+/// Exciting documentation
+/// Maybe even multiple lines
+type Wibble {
+    /// Some more exciting documentation
+    Wibble
+}
+",
+        find_position_of("Wibble").nth_occurrence(2)
+    );
+}
+
+#[test]
+fn hover_type_constructor_with_label() {
+    assert_hover!(
+        "
+/// Exciting documentation
+/// Maybe even multiple lines
+type Wibble {
+    /// Some more exciting documentation
+    Wibble(arg: String)
+    /// The most exciting documentation
+    Wobble(Int)
+}
+",
+        find_position_of("Wibble").nth_occurrence(2)
+    );
+}
+
+#[test]
+fn hover_type_constructor_with_no_label() {
+    assert_hover!(
+        "
+/// Exciting documentation
+/// Maybe even multiple lines
+type Wibble {
+    /// Some more exciting documentation
+    Wibble(arg: String)
+    /// The most exciting documentation
+    Wobble(Int)
+}
+",
+        find_position_of("Wobble")
+    );
+}